Minor improvements to pip_update.sh
This commit includes some syntactic changes, a better expression to make sure
the script works no matter what directory it's called from, a trap function to
make sure cleanup of the temporary virtualenv we create happens, and a test that
/etc/os-release exists before trying to grep through it.

+++ pip_update.sh
@@ -4,13 +4,15 @@
 # Run periodically to keep Python requirements up-to-date
 set -e
 
-REQUIREMENTS_DIR=$(pwd)/securedrop/requirements
+dir="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
+requirements_dir="${dir}/securedrop/requirements"
+venv="review_env"
 
 # This script should not be run with an active virtualenv. Calling deactivate
 # does not work reliably, so instead we warn then quit.
 if [[ -n $VIRTUAL_ENV ]]; then
   echo "Please deactivate your virtualenv before running this script."
-  exit
+  exit 1
 fi
 
 # Test if pip and virtualenv are available and install them if not
@@ -20,7 +22,7 @@ command -v virtualenv > /dev/null
 virualenv_installed=$?
 
 if [[ $pip_installed -ne 0 ]] || [[ $virtualenv_installed -ne 0 ]]; then
-  if $(grep -i "debian" /etc/os-release > /dev/null); then
+  if [[ -e /etc/os-release ]] && $(grep -i "debian" /etc/os-release > /dev/null); then
     sudo apt-get install -y python-pip virtualenv
   else
     echo "This script requires pip and virtualenv to run."
@@ -30,24 +32,22 @@ fi
 
 # Create a temporary virtualenv for the SecureDrop Python packages in our
 # requirements directory
-cd $REQUIREMENTS_DIR
-VENV=review_env
-virtualenv -p python2.7 $VENV
-source $VENV/bin/activate
+cd $requirements_dir
+
+trap "rm -rf ${venv}" EXIT
+
+virtualenv -p python2.7 $venv
+source "${venv}/bin/activate"
 
 # Install the most recent pip that pip-tools supports and the latest pip-tools
 # (must be done in order as the former is a dependency of the latter).
-pip install --upgrade pip==8.1.1
+pip install pip==8.1.1
 pip install pip-tools
 
 # Compile new requirements (.txt) files from our top-level dependency (.in)
 # files. See http://nvie.com/posts/better-package-management/
-for r in "securedrop-requirements" "test-requirements"
-do
+for r in "securedrop" "test"; do
   # Maybe pip-tools will get its act together and standardize their cert-pinning
   # syntax and this line will break. One can only hope.
-  pip-compile -o $r".txt" $r".in"
+  pip-compile -o "${r}-requirements.txt" "${r}-requirements.in"
 done
-
-# Remove the temporary virtualenv
-rm -r $VENV
